Bypass dispatchInterval's setter in ReplayQueue's init due to ReplayIO-ReplayQueue circular init dependency cycle
This is kind of a hack. ReplayIO's init method instantiates ReplayQueue, which in turn originally called the dispatchInterval setter method. However, the use of DEBUG_LOG in the setter method caused the program to stall, because DEBUG_LOG references ReplayIO's debugMode property, and at this point in time ReplayIO hasn't finished initializing itself. Thus, we're stuck in a circular dependency loop.

Before refactoring ReplayQueue to not be a singleton, this wasn't an issue, as ReplayQueue was lazily loaded *after* ReplayIO was finished being initialized.

Even though singletons weren't causing any real problems before, it's better to avoid them if possible (potentially making testing easier). Whew that was long!